+++ src/main/java/echo/command/DeadlineCommand.java
@@ -1,3 +1,11 @@
+package echo.command;
+
+import echo.storage.Storage;
+import echo.task.DeadlineTask;
+import echo.task.TaskList;
+import echo.ui.Ui;
+import echo.main.EchoException;
+
 import java.io.IOException;
 import java.time.LocalDateTime;
