+++ src/main/java/echo/command/DeleteCommand.java
@@ -1,19 +1,27 @@
+package echo.command;
+
+import echo.storage.Storage;
+import echo.task.TaskList;
+import echo.ui.Ui;
+import echo.main.EchoException;
+
 import java.io.IOException;
 
-public class UnmarkCommand extends Command {
+public class DeleteCommand extends Command {
 
-    public static final String COMMAND = "unmark";
+    public static final String COMMAND = "delete";
 
     private final int TASK_INDEX;
 
-    public UnmarkCommand(int taskIndex) {
+    public DeleteCommand(int taskIndex) {
         this.TASK_INDEX = taskIndex;
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws EchoException {
         try {
-            tasks.unmark(TASK_INDEX);
-            ui.showUnmark(tasks.taskStatus(TASK_INDEX));
+            String taskStatus = tasks.taskStatus(TASK_INDEX);
+            tasks.delete(TASK_INDEX);
+            ui.showDelete(taskStatus, tasks.size());
             storage.save(tasks);
         } catch (IndexOutOfBoundsException e) {
             // Task number provided by user does not exist in tasks.
