+++ src/main/java/echo/command/UnmarkCommand.java
@@ -1,20 +1,26 @@
+package echo.command;
+
+import echo.storage.Storage;
+import echo.task.TaskList;
+import echo.ui.Ui;
+import echo.main.EchoException;
+
 import java.io.IOException;
 
-public class DeleteCommand extends Command {
+public class UnmarkCommand extends Command {
 
-    public static final String COMMAND = "delete";
+    public static final String COMMAND = "unmark";
 
     private final int TASK_INDEX;
 
-    public DeleteCommand(int taskIndex) {
+    public UnmarkCommand(int taskIndex) {
         this.TASK_INDEX = taskIndex;
     }
 
     public void execute(TaskList tasks, Ui ui, Storage storage) throws EchoException {
         try {
-            String taskStatus = tasks.taskStatus(TASK_INDEX);
-            tasks.delete(TASK_INDEX);
-            ui.showDelete(taskStatus, tasks.size());
+            tasks.unmark(TASK_INDEX);
+            ui.showUnmark(tasks.taskStatus(TASK_INDEX));
             storage.save(tasks);
         } catch (IndexOutOfBoundsException e) {
             // Task number provided by user does not exist in tasks.
