+++ src/main/java/echo/task/EventTask.java
@@ -1,3 +1,5 @@
+package echo.task;
+
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
