+++ src/main/java/echo/task/Task.java
@@ -1,3 +1,5 @@
+package echo.task;
+
 /**
  * Encapsulates a task.
  */
