+++ src/main/java/echo/task/TodoTask.java
@@ -1,3 +1,5 @@
+package echo.task;
+
 /**
  * TodoTask task which inherits from Task class.
  */
