+++ src/main/java/echo/ui/Ui.java
@@ -1,3 +1,5 @@
+package echo.ui;
+
 import java.util.Scanner;
 
 /**
