+++ text-ui-test/EXPECTED.TXT
@@ -5,61 +5,87 @@ U _____ u    ____    _   _      U  ___ u
  |_____|    \____|  |_| |_|   \_)-\___/  
  <<   >>   _// \\   //   \\        \\    
 (__) (__) (__)(__) (_") ("_)      (__) 
-        ____________________________________________________________
-        Got it. I've added this task: 
-          4. [T][ ] borrow book
-        Now you have 4 tasks in the list.
-        ____________________________________________________________
-        ____________________________________________________________
-        Got it. I've added this task: 
-          5. [D][ ] sell book (by: Sunday)
-        Now you have 5 tasks in the list.
-        ____________________________________________________________
-        ____________________________________________________________
-        Got it. I've added this task: 
-          6. [E][ ] get rich (at: Mon 2-4pm)
-        Now you have 6 tasks in the list.
-        ____________________________________________________________
-        ____________________________________________________________
-        1. [T][X] read book 
-        2. [D][ ] return book (by: June 6th) 
-        3. [E][ ] project meeting (at: Aug 6th 2-4pm) 
-        4. [T][ ] borrow book 
-        5. [D][ ] sell book (by: Sunday) 
-        6. [E][ ] get rich (at: Mon 2-4pm)
-        ____________________________________________________________
-        ____________________________________________________________
-        Nice! The task is marked as done: 
-          4. [T][X] borrow book
-        ____________________________________________________________
-        ____________________________________________________________
-        1. [T][X] read book 
-        2. [D][ ] return book (by: June 6th) 
-        3. [E][ ] project meeting (at: Aug 6th 2-4pm) 
-        4. [T][X] borrow book 
-        5. [D][ ] sell book (by: Sunday) 
-        6. [E][ ] get rich (at: Mon 2-4pm)
-        ____________________________________________________________
-        ____________________________________________________________
-        Noted. I've removed the task: 
-          4. [T][X] borrow book 
-        Now you have 5 tasks in the list.
-        ____________________________________________________________
-        ____________________________________________________________
-        Noted. I've removed the task: 
-          4. [D][ ] sell book (by: Sunday) 
-        Now you have 4 tasks in the list.
-        ____________________________________________________________
-        ____________________________________________________________
-        Noted. I've removed the task: 
-          4. [E][ ] get rich (at: Mon 2-4pm) 
-        Now you have 3 tasks in the list.
-        ____________________________________________________________
-        ____________________________________________________________
-        1. [T][X] read book 
-        2. [D][ ] return book (by: June 6th) 
-        3. [E][ ] project meeting (at: Aug 6th 2-4pm)
-        ____________________________________________________________
-        ____________________________________________________________
-        Goodbye!
-        ____________________________________________________________
+
+    ____________________________________________________________
+    Commands: 
+    list                                | List current tasks. 
+    todo <description>                  | Add a todo task. 
+    deadline <description> /by <time>   | Add a deadline task to be done before time. 
+    event <description> /at <time>      | Add an event task that occurs at time. 
+    mark <task number>                  | Mark task as completed. 
+    unmark <task number>                | Unmark task as uncompleted. 
+    delete <task number>                | Delete task. 
+    bye                                 | exit. 
+    ____________________________________________________________
+    ____________________________________________________________
+    Got it. I've added this task: 
+      4. [T][ ] borrow book
+    Now you have 4 tasks in the list.
+    ____________________________________________________________
+    ____________________________________________________________
+    Invalid format for date and time. Follow the <yyyy-mm-dd> <24hr time> format.
+    E.g. 2019-10-15 1800
+    ____________________________________________________________
+    ____________________________________________________________
+    Invalid input. Make sure to include '/by'
+    E.g. deadline return book /by 2019-10-15 1800
+    ____________________________________________________________
+    ____________________________________________________________
+    Got it. I've added this task: 
+      5. [D][ ] sell book (by: Jan 01 2022, 6:00PM)
+    Now you have 5 tasks in the list.
+    ____________________________________________________________
+    ____________________________________________________________
+    Invalid input. Make sure to include '/at' 
+.    E.g. event meeting /at 2019-10-15 1800
+    ____________________________________________________________
+    ____________________________________________________________
+    Got it. I've added this task: 
+      6. [E][ ] get rich (at: Apr 20 2022, 4:00AM)
+    Now you have 6 tasks in the list.
+    ____________________________________________________________
+    ____________________________________________________________
+    1. [T][X] read book
+    2. [D][ ] return book (by: Jan 01 2022, 6:00PM)
+    3. [E][ ] project meeting (at: Feb 04 2022, 4:00AM)
+    4. [T][ ] borrow book
+    5. [D][ ] sell book (by: Jan 01 2022, 6:00PM)
+    6. [E][ ] get rich (at: Apr 20 2022, 4:00AM)
+    ____________________________________________________________
+    ____________________________________________________________
+    Nice! The task is marked as done: 
+      4. [T][X] borrow book
+    ____________________________________________________________
+    ____________________________________________________________
+    1. [T][X] read book
+    2. [D][ ] return book (by: Jan 01 2022, 6:00PM)
+    3. [E][ ] project meeting (at: Feb 04 2022, 4:00AM)
+    4. [T][X] borrow book
+    5. [D][ ] sell book (by: Jan 01 2022, 6:00PM)
+    6. [E][ ] get rich (at: Apr 20 2022, 4:00AM)
+    ____________________________________________________________
+    ____________________________________________________________
+    OK! The task is unmarked: 
+      4. [T][ ] borrow book
+    ____________________________________________________________
+    ____________________________________________________________
+    Noted. I've removed the task: 
+      4. [T][ ] borrow book
+    Now you have 5 tasks in the list.
+    ____________________________________________________________
+    ____________________________________________________________
+    Noted. I've removed the task: 
+      4. [D][ ] sell book (by: Jan 01 2022, 6:00PM)
+    Now you have 4 tasks in the list.
+    ____________________________________________________________
+    ____________________________________________________________
+    Noted. I've removed the task: 
+      4. [E][ ] get rich (at: Apr 20 2022, 4:00AM)
+    Now you have 3 tasks in the list.
+    ____________________________________________________________
+    ____________________________________________________________
+    1. [T][X] read book
+    2. [D][ ] return book (by: Jan 01 2022, 6:00PM)
+    3. [E][ ] project meeting (at: Feb 04 2022, 4:00AM)
+    ____________________________________________________________
+    Goodbye!
